// frontend/src/core/constants/types.ts
export enum AuthEnums {
  SET_USER = "myoceanmath/auth/SET_USER",
  SET_AUTH = "myoceanmath/auth/SET_AUTH",
  SET_USER_TOKEN = "myoceanmath/auth/SET_USER_TOKEN",
  SET_LOADER = "myoceanmath/auth/SET_LOADER",
  IS_NOT_USER = "myoceanmath/auth/IS_NOT_USER",
  LOGOUT_USER = "myoceanmath/auth/LOGOUT_USER",
  SET_LOGIN_ERROR = "myoceanmath/auth/SET_LOGIN_ERROR",
  SET_FINISHED_TASKS = "myoceanmath/auth/SET_FINISHED_TASKS",
  SET_LOADER_TASKS = "myoceanmath/auth/SET_LOADER_TASKS"
}

export enum AppEnums {
  SET_MODAL = "myoceanmath/app/SET_MODAL"
}

export enum LevelEnums {
  GET_TASKS = "myoceanmath/level/GET_TASKS",
  SET_LEVEL = "myoceanmath/level/SET_LEVEL",
  SET_CURRENT_TASK = "myoceanmath/level/SET_CURRENT_TASK",
  SET_FINISHED = "myoceanmath/level/SET_FINISHED",
  CLEAR_TASKS = "myoceanmath/level/CLEAR_TASKS",
  SET_ERROR = "myoceanmath/level/SET_ERROR",
  GET_ERRORS = "myoceanmath/level/GET_ERRORS"
}